# spec/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Api Documentation",
    "version" : "1.0",
    "title" : "Api Documentation",
    "termsOfService" : "urn:tos",
    "contact" : { },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host" : "localhost",
  "basePath" : "/",
...
  "paths" : {
...
    "/api/discount" : {
      "post" : {
        "tags" : [ "discount-controller" ],
        "summary" : "createDiscount",
        "operationId" : "createDiscountUsingPOST",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "discount",
          "description" : "discount",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Discount"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Discount"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/discount/batch" : {
      "post" : {
        "tags" : [ "discount-controller" ],
        "summary" : "createDiscounts",
        "operationId" : "createDiscountsUsingPOST",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "discounts",
          "description" : "discounts",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/Discount"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Discount"
              }
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/discount/in-list/{salesOrg}" : {
      "get" : {
        "tags" : [ "discount-controller" ],
        "summary" : "getAllDiscountsForSalesOrgAndMaterialNumbersInList",
        "operationId" : "getAllDiscountsForSalesOrgAndMaterialNumbersInListUsingGET",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "materialNumbers",
          "in" : "query",
          "description" : "materialNumbers",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesOffice",
          "in" : "query",
          "description" : "salesOffice",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesOrg",
          "in" : "path",
          "description" : "salesOrg",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "zones",
          "in" : "query",
          "description" : "zones",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Discount"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/api/discount/list" : {
      "get" : {
        "tags" : [ "discount-controller" ],
        "summary" : "getAllDiscounts",
        "operationId" : "getAllDiscountsUsingGET",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Discount"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/discount/list/{salesOrg}" : {
      "get" : {
        "tags" : [ "discount-controller" ],
        "summary" : "getAllDiscountsForSalesOrg",
        "operationId" : "getAllDiscountsForSalesOrgUsingGET",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "materialNumber",
          "in" : "query",
          "description" : "materialNumber",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "salesOffice",
          "in" : "query",
          "description" : "salesOffice",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesOrg",
          "in" : "path",
          "description" : "salesOrg",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "zone",
          "in" : "query",
          "description" : "zone",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Discount"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/discount/{id}" : {
      "put" : {
        "tags" : [ "discount-controller" ],
        "summary" : "updateDiscount",
        "operationId" : "updateDiscountUsingPUT",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "discount",
          "description" : "discount",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Discount"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Discount"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/api/v1/discount" : {
      "post" : {
        "tags" : [ "discount-controller" ],
        "summary" : "createDiscount",
        "operationId" : "createDiscountUsingPOST_1",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "discount",
          "description" : "discount",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Discount"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Discount"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/discount/batch" : {
      "post" : {
        "tags" : [ "discount-controller" ],
        "summary" : "createDiscounts",
        "operationId" : "createDiscountsUsingPOST_1",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "discounts",
          "description" : "discounts",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/Discount"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Discount"
              }
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/discount/in-list/{salesOrg}" : {
      "get" : {
        "tags" : [ "discount-controller" ],
        "summary" : "getAllDiscountsForSalesOrgAndMaterialNumbersInList",
        "operationId" : "getAllDiscountsForSalesOrgAndMaterialNumbersInListUsingGET_1",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "materialNumbers",
          "in" : "query",
          "description" : "materialNumbers",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesOffice",
          "in" : "query",
          "description" : "salesOffice",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesOrg",
          "in" : "path",
          "description" : "salesOrg",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "zones",
          "in" : "query",
          "description" : "zones",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Discount"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/api/v1/discount/list" : {
      "get" : {
        "tags" : [ "discount-controller" ],
        "summary" : "getAllDiscounts",
        "operationId" : "getAllDiscountsUsingGET_1",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Discount"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/discount/list/{salesOrg}" : {
      "get" : {
        "tags" : [ "discount-controller" ],
        "summary" : "getAllDiscountsForSalesOrg",
        "operationId" : "getAllDiscountsForSalesOrgUsingGET_1",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "materialNumber",
          "in" : "query",
          "description" : "materialNumber",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "salesOffice",
          "in" : "query",
          "description" : "salesOffice",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "salesOrg",
          "in" : "path",
          "description" : "salesOrg",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "zone",
          "in" : "query",
          "description" : "zone",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Discount"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
    "/api/v1/discount/{id}" : {
      "put" : {
        "tags" : [ "discount-controller" ],
        "summary" : "updateDiscount",
        "operationId" : "updateDiscountUsingPUT_1",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "discount",
          "description" : "discount",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Discount"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Discount"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Discount" : {
      "type" : "object",
      "properties" : {
        "deviceType" : {
          "type" : "string"
        },
        "discountLevels" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DiscountLevel"
          }
        },
        "fa" : {
          "type" : "string"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "materialDesignation" : {
          "type" : "string"
        },
        "materialNumber" : {
          "type" : "string"
        },
        "salesOffice" : {
          "type" : "string"
        },
        "salesOrg" : {
          "type" : "string"
        },
        "standardPrice" : {
          "type" : "number",
          "format" : "double"
        }
      },
      "title" : "Discount"
    },
    "DiscountLevel" : {
      "type" : "object",
      "properties" : {
        "calculatedDiscount" : {
          "type" : "number",
          "format" : "double"
        },
        "discount" : {
          "type" : "number",
          "format" : "double"
        },
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "level" : {
          "type" : "integer",
          "format" : "int32"
        },
        "pctDiscount" : {
          "type" : "number",
          "format" : "double"
        },
        "zone" : {
          "type" : "integer",
          "format" : "int32"
        }
      },
      "title" : "DiscountLevel"
    },
...
  }
}